#108 get_hash returns unsigned

+++ hashtable/hashtable.c
@@ -31,12 +31,12 @@ static int hash_function(const struct YASL_Object s, const int a, const int m) {
         }
 }
 
-static int get_hash(const struct YASL_Object s, const int num_buckets, const int attempt) {
+static unsigned int get_hash(const struct YASL_Object s, const int num_buckets, const int attempt) {
         const int hash_a = hash_function(s, PRIME_A, num_buckets);
         if (attempt == 0)
                 return hash_a % num_buckets;
         const int hash_b = hash_function(s, PRIME_B, num_buckets);
-        return (hash_a + (attempt * (hash_b + 1))) % num_buckets;
+        return ((unsigned int)(hash_a + (attempt * (hash_b + 1)))) % num_buckets;
 }
 
 static Item_t new_item(const struct YASL_Object k, const struct YASL_Object v) {
